+++ ikii-back-commons/src/main/java/mx/ikii/commons/payload/response/product/ProductBusinessResponse.java
@@ -17,5 +17,6 @@ public class ProductBusinessResponse {
 	private Map<String, String> fullDescription; // for Food
 	private BigDecimal price;
 	private String pathImage;
+	private boolean favorite;
 
 }