+++ ikii-back-microservices/ikii-back-customers/src/main/java/mx/ikii/customers/controller/ICustomerDetailsController.java
@@ -19,22 +19,25 @@ public interface ICustomerDetailsController {
 
 	@GetMapping("/{id}")
 	ResponseEntity<CustomerDetailsResponse> getById(String id);
-	
+
 	@GetMapping("/customers/{customerId}")
 	ResponseEntity<CustomerDetailsResponse> getByCustomerId(String customerId);
-	
+
 	@GetMapping
 	ResponseEntity<Page<CustomerDetailsResponse>> getAll(Pageable pageable);
-	
+
 	@PostMapping
 	ResponseEntity<CustomerDetailsResponse> create(CustomerDetailsRequest customerDetailsRequest);
-	
+
 	@PutMapping("/{id}")
 	ResponseEntity<CustomerDetailsResponse> update(CustomerDetailsRequest customerDetailsRequest, String id);
-	
+
 	@PutMapping("/customers/{customerId}/business-favorites/{businessId}")
 	ResponseEntity<Void> toggleBusinessFavorite(String customerId, String businessId);
-	
+
+	@PutMapping("/customers/{customerId}/product-favorites/{productId}")
+	ResponseEntity<Void> toggleProductFavorite(String customerId, String productId);
+
 	@DeleteMapping("/{id}")
 	ResponseEntity<Void> delete(String id);